Resolved first review code wave
