+++ .gitignore
@@ -3,8 +3,6 @@
 __pycache__/
 *.py[cod]
 
-# db
-*.sqlite3
 
 # C extensions
 *.so
