+++ src/helperModules/definitions.py
@@ -1,0 +1,5 @@
+from dotenv import dotenv_values
+
+config = dotenv_values('.env')
+
+TOKEN = config.get('TOKEN')
