+++ src/helperModules/string_helper.py
@@ -1,0 +1,14 @@
+import re
+
+
+class StringHelper:
+    URL_PATTERN = (
+        r'http[s]?://'  # Scheme
+        r'(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\\(\\),]'
+        r'|(?:%[0-9a-fA-F][0-9a-fA-F]))+'
+    )
+
+    @staticmethod
+    def is_url(message_content):
+        return re.match(StringHelper.URL_PATTERN, 
+                        message_content)
